Factorized code for assigning users to sites
